refactor: ProfileHeaderCardView 및 ProfileStatItemView 분리를 위한 구조 정리
- ProfileView에서 컴포넌트 호출 구조 유지
